work on UI error feedback

+++ src/ui/profile-page.mjs/delete-listing.mjs
@@ -1,15 +1,19 @@
 import { deleteListing } from '../../api-calls/listings/delete-listing.mjs';
 
+/**
+ * Sets up deletebutton, corrosponding with specific id of the listing
+ * @param {string} container - Takes the container for the listing
+ */
 export async function setupDeleteListingInteractions(container) {
   container.addEventListener('click', async function (e) {
     const deleteBtn = e.target.closest('.delete-button');
     if (deleteBtn) {
-      const listingId = deleteBtn.getAttribute('data-listing-id');
-      try {
+      const isConfirmed = confirm(
+        'Are you sure you want to delete this listing? This action cannot be undone.'
+      );
+      if (isConfirmed) {
+        const listingId = deleteBtn.getAttribute('data-listing-id');
         await deleteListing(listingId);
-        window.location.reload();
-      } catch (error) {
-        console.log('wont delete');
       }
     }
   });
